Mve webgpu to a thin Approach so we can import the correct imports
