+++ packages/dev/core/src/Engines/WebGPU/Extensions/engine.readTexture.ts
@@ -1,6 +1,6 @@
+import { ThinWebGPUEngine } from "core/Engines/thinWebGPUEngine";
 import type { InternalTexture } from "../../../Materials/Textures/internalTexture";
 import type { Nullable } from "../../../types";
-import { WebGPUEngine } from "../../webgpuEngine";
 import type { WebGPUHardwareTexture } from "../webgpuHardwareTexture";
 
 declare module "../../abstractEngine" {
@@ -35,7 +35,7 @@ declare module "../../abstractEngine" {
     }
 }
 
-WebGPUEngine.prototype._readTexturePixels = function (
+ThinWebGPUEngine.prototype._readTexturePixels = function (
     texture: InternalTexture,
     width: number,
     height: number,
@@ -56,7 +56,7 @@ WebGPUEngine.prototype._readTexturePixels = function (
     return this._textureHelper.readPixels(gpuTextureWrapper.underlyingResource!, x, y, width, height, gpuTextureWrapper.format, faceIndex, level, buffer, noDataConversion);
 };
 
-WebGPUEngine.prototype._readTexturePixelsSync = function (): ArrayBufferView {
+ThinWebGPUEngine.prototype._readTexturePixelsSync = function (): ArrayBufferView {
     // eslint-disable-next-line no-throw-literal
     throw "_readTexturePixelsSync is unsupported in WebGPU!";
 };
