+++ packages/dev/core/src/Engines/WebGPU/Extensions/engine.renderTargetTexture.ts
@@ -1,6 +1,6 @@
 import type { Nullable } from "../../../types";
-import { WebGPUEngine } from "../../webgpuEngine";
 import type { RenderTargetTexture } from "../../../Materials/Textures/renderTargetTexture";
+import { ThinWebGPUEngine } from "core/Engines/thinWebGPUEngine";
 
 declare module "../../abstractEngine" {
     export interface AbstractEngine {
@@ -15,7 +15,12 @@ declare module "../../abstractEngine" {
     }
 }
 
-WebGPUEngine.prototype.setDepthStencilTexture = function (channel: number, uniform: Nullable<WebGLUniformLocation>, texture: Nullable<RenderTargetTexture>, name?: string): void {
+ThinWebGPUEngine.prototype.setDepthStencilTexture = function (
+    channel: number,
+    uniform: Nullable<WebGLUniformLocation>,
+    texture: Nullable<RenderTargetTexture>,
+    name?: string
+): void {
     if (!texture || !texture.depthStencilTexture) {
         this._setTexture(channel, null, undefined, undefined, name);
     } else {
