+++ packages/dev/core/src/Engines/webgpuEngine.ts
@@ -1,5 +1,6 @@
 /* eslint-disable babylonjs/available */
 import { Logger } from "../Misc/logger";
+import { ThinWebGPUEngine } from "./thinWebGPUEngine";
 import type { Nullable, DataArray, IndicesArray, Immutable, FloatArray } from "../types";
 import { Color4 } from "../Maths/math";
 import { InternalTexture, InternalTextureSource } from "../Materials/Textures/internalTexture";
@@ -31,7 +32,6 @@ import { WebGPUHardwareTexture } from "./WebGPU/webgpuHardwareTexture";
 import type { IColor4Like } from "../Maths/math.like";
 import { UniformBuffer } from "../Materials/uniformBuffer";
 import { WebGPUCacheSampler } from "./WebGPU/webgpuCacheSampler";
-import type { WebGPUCacheRenderPipeline } from "./WebGPU/webgpuCacheRenderPipeline";
 import { WebGPUCacheRenderPipelineTree } from "./WebGPU/webgpuCacheRenderPipelineTree";
 import { WebGPUStencilStateComposer } from "./WebGPU/webgpuStencilStateComposer";
 import { WebGPUDepthCullingState } from "./WebGPU/webgpuDepthCullingState";
@@ -61,7 +61,6 @@ import "../Buffers/buffer.align";
 import type { VideoTexture } from "../Materials/Textures/videoTexture";
 import type { RenderTargetTexture } from "../Materials/Textures/renderTargetTexture";
 import type { RenderTargetWrapper } from "./renderTargetWrapper";
-import { WebGPUPerfCounter } from "./WebGPU/webgpuPerfCounter";
 import type { Scene } from "../scene";
 
 import { SphericalPolynomial } from "../Maths/sphericalPolynomial";
@@ -89,6 +88,13 @@ import { resetCachedPipeline } from "../Materials/effect.functions";
 import { WebGPUExternalTexture } from "./WebGPU/webgpuExternalTexture";
 import type { TextureSampler } from "../Materials/Textures/textureSampler";
 import type { StorageBuffer } from "../Buffers/storageBuffer";
+import "./WebGPU/Extensions/engine.alpha";
+import "./WebGPU/Extensions/engine.rawTexture";
+import "./WebGPU/Extensions/engine.readTexture";
+import "./WebGPU/Extensions/engine.cubeTexture";
+import "./WebGPU/Extensions/engine.renderTarget";
+import "./WebGPU/Extensions/engine.renderTargetTexture";
+import "./WebGPU/Extensions/engine.renderTargetCube";
 
 const viewDescriptorSwapChainAntialiasing: GPUTextureViewDescriptor = {
     label: `TextureView_SwapChain_ResolveTarget`,
@@ -194,7 +200,7 @@ export interface WebGPUEngineOptions extends AbstractEngineOptions, GPURequestAd
  * The web GPU engine class provides support for WebGPU version of babylon.js.
  * @since 5.0.0
  */
-export class WebGPUEngine extends AbstractEngine {
+export class WebGPUEngine extends ThinWebGPUEngine {
     // Default glslang options.
     private static readonly _GlslangDefaultOptions: GlslangOptions = {
         jsPath: `${Tools._DefaultCdnUrl}/glslang/glslang.js`,
@@ -241,22 +247,14 @@ export class WebGPUEngine extends AbstractEngine {
     private _glslangOptions?: GlslangOptions;
     private _twgslOptions?: TwgslOptions;
     /** @internal */
-    public _textureHelper: WebGPUTextureManager;
-    /** @internal */
     public _bufferManager: WebGPUBufferManager;
     private _clearQuad: WebGPUClearQuad;
     /** @internal */
     public _cacheSampler: WebGPUCacheSampler;
-    /** @internal */
-    public _cacheRenderPipeline: WebGPUCacheRenderPipeline;
     private _cacheBindGroups: WebGPUCacheBindGroups;
     private _emptyVertexBuffer: VertexBuffer;
     /** @internal */
     public _mrtAttachments: number[];
-    /** @internal */
-    public _timestampQuery: WebGPUTimestampQuery;
-    /** @internal */
-    public _timestampIndex = 0;
     /** @internal */
     public _occlusionQuery: WebGPUOcclusionQuery;
     /** @internal */
@@ -313,17 +311,10 @@ export class WebGPUEngine extends AbstractEngine {
     /** @internal */
     public _ubDontInvertY: WebGPUDataBuffer;
 
-    // Frame Life Cycle (recreated each frame)
-    /** @internal */
-    public _uploadEncoder: GPUCommandEncoder;
-    /** @internal */
-    public _renderEncoder: GPUCommandEncoder;
-
     private _commandBuffers: GPUCommandBuffer[] = [null as any, null as any];
 
     // Frame Buffer Life Cycle (recreated for each render target pass)
-    /** @internal */
-    public _currentRenderPass: Nullable<GPURenderPassEncoder> = null;
+
     private _mainRenderPassWrapper: IWebGPURenderPassWrapper = {
         renderPassDescriptor: null,
         colorAttachmentViewDescriptor: null,
@@ -340,11 +331,6 @@ export class WebGPUEngine extends AbstractEngine {
     };
     /** @internal */
     public _pendingDebugCommands: Array<[string, Nullable<string>]> = [];
-    /**
-     * Used for both the compatibilityMode=false and the snapshot rendering modes (as both can't be enabled at the same time)
-     * @internal
-     */
-    public _bundleList: WebGPUBundleList;
 
     // DrawCall Life Cycle
     // Effect is on the parent class
@@ -359,23 +345,6 @@ export class WebGPUEngine extends AbstractEngine {
     private _currentIndexBuffer: Nullable<DataBuffer> = null;
     private _colorWriteLocal = true;
     private _forceEnableEffect = false;
-
-    // TODO WEBGPU remove those variables when code stabilized
-    /** @internal */
-    public dbgShowShaderCode = false;
-    /** @internal */
-    public dbgSanityChecks = true;
-    /** @internal */
-    public dbgVerboseLogsForFirstFrames = false;
-    /** @internal */
-    public dbgVerboseLogsNumFrames = 10;
-    /** @internal */
-    public dbgLogIfNotDrawWrapper = true;
-    /** @internal */
-    public dbgShowEmptyEnableEffectCalls = true;
-
-    private _snapshotRendering: WebGPUSnapshotRendering;
-    protected _snapshotRenderingMode = Constants.SNAPSHOTRENDERING_STANDARD;
 
     /**
      * Gets or sets the snapshot rendering mode
@@ -553,29 +522,6 @@ export class WebGPUEngine extends AbstractEngine {
     public override set compatibilityMode(mode: boolean) {
         this._compatibilityMode = mode;
     }
-
-    /**
-     * Enables or disables GPU timing measurements.
-     * Note that this is only supported if the "timestamp-query" extension is enabled in the options.
-     */
-    public get enableGPUTimingMeasurements(): boolean {
-        return this._timestampQuery.enable;
-    }
-
-    public set enableGPUTimingMeasurements(enable: boolean) {
-        if (this._timestampQuery.enable === enable) {
-            return;
-        }
-        (this.gpuTimeInFrameForMainPass as any) = enable ? new WebGPUPerfCounter() : undefined;
-        this._timestampQuery.enable = enable;
-    }
-
-    /**
-     * Gets the GPU time spent in the main render pass for the last frame rendered (in nanoseconds).
-     * You have to enable the "timestamp-query" extension in the engine constructor options and set engine.enableGPUTimingMeasurements = true.
-     * It will only return time spent in the main pass, not additional render target / compute passes (if any)!
-     */
-    public readonly gpuTimeInFrameForMainPass?: WebGPUPerfCounter;
 
     /** @internal */
     public get currentSampleCount(): number {
@@ -1264,10 +1210,6 @@ export class WebGPUEngine extends AbstractEngine {
      */
     public _getShaderProcessingContext(shaderLanguage: ShaderLanguage, pureMode: boolean): Nullable<ShaderProcessingContext> {
         return new WebGPUShaderProcessingContext(shaderLanguage, pureMode);
-    }
-
-    private _currentPassIsMainPass() {
-        return this._currentRenderTarget === null;
     }
 
     private _getCurrentRenderPass(): GPURenderPassEncoder {
@@ -2806,53 +2748,6 @@ export class WebGPUEngine extends AbstractEngine {
     }
 
     /**
-     * @internal
-     */
-    public _generateMipmaps(texture: InternalTexture, commandEncoder?: GPUCommandEncoder) {
-        commandEncoder = commandEncoder ?? this._renderEncoder;
-
-        const gpuHardwareTexture = texture._hardwareTexture as Nullable<WebGPUHardwareTexture>;
-
-        if (!gpuHardwareTexture) {
-            return;
-        }
-
-        if (commandEncoder === this._renderEncoder) {
-            // We must close the current pass (if any) because we are going to use the render encoder to generate the mipmaps (so, we are going to create a new render pass)
-            this._endCurrentRenderPass();
-        }
-
-        const format = (texture._hardwareTexture as WebGPUHardwareTexture).format;
-        const mipmapCount = WebGPUTextureHelper.ComputeNumMipmapLevels(texture.width, texture.height);
-
-        if (this.dbgVerboseLogsForFirstFrames) {
-            if ((this as any)._count === undefined) {
-                (this as any)._count = 0;
-            }
-            if (!(this as any)._count || (this as any)._count < this.dbgVerboseLogsNumFrames) {
-                Logger.Log(
-                    "frame #" +
-                        (this as any)._count +
-                        " - generate mipmaps - width=" +
-                        texture.width +
-                        ", height=" +
-                        texture.height +
-                        ", isCube=" +
-                        texture.isCube +
-                        ", command encoder=" +
-                        (commandEncoder === this._renderEncoder ? "render" : "copy")
-                );
-            }
-        }
-
-        if (texture.isCube) {
-            this._textureHelper.generateCubeMipmaps(gpuHardwareTexture, format, mipmapCount, commandEncoder);
-        } else {
-            this._textureHelper.generateMipmaps(gpuHardwareTexture, format, mipmapCount, 0, texture.is3D, commandEncoder);
-        }
-    }
-
-    /**
      * Update a portion of an internal texture
      * @param texture defines the texture to update
      * @param imageData defines the data to store into the texture
@@ -3371,49 +3266,6 @@ export class WebGPUEngine extends AbstractEngine {
         if (!this._isStencilEnable) {
             this._stencilStateComposer.enabled = false;
         }
-    }
-
-    /** @internal */
-    public _endCurrentRenderPass(): number {
-        if (!this._currentRenderPass) {
-            return 0;
-        }
-
-        const currentPassIndex = this._currentPassIsMainPass() ? 2 : 1;
-
-        if (!this._snapshotRendering.endRenderPass(this._currentRenderPass) && !this.compatibilityMode) {
-            this._bundleList.run(this._currentRenderPass);
-            this._bundleList.reset();
-        }
-        this._currentRenderPass.end();
-
-        this._timestampQuery.endPass(
-            this._timestampIndex,
-            (this._currentRenderTarget && (this._currentRenderTarget as WebGPURenderTargetWrapper).gpuTimeInFrame
-                ? (this._currentRenderTarget as WebGPURenderTargetWrapper).gpuTimeInFrame
-                : this.gpuTimeInFrameForMainPass) as WebGPUPerfCounter
-        );
-        this._timestampIndex += 2;
-
-        if (this.dbgVerboseLogsForFirstFrames) {
-            if ((this as any)._count === undefined) {
-                (this as any)._count = 0;
-            }
-            if (!(this as any)._count || (this as any)._count < this.dbgVerboseLogsNumFrames) {
-                Logger.Log(
-                    "frame #" +
-                        (this as any)._count +
-                        " - " +
-                        (currentPassIndex === 2 ? "main" : "render target") +
-                        " end pass" +
-                        (currentPassIndex === 1 ? " - internalTexture.uniqueId=" + this._currentRenderTarget?.texture?.uniqueId : "")
-                );
-            }
-        }
-        this._debugPopGroup?.(0);
-        this._currentRenderPass = null;
-
-        return currentPassIndex;
     }
 
     /**
