+++ packages/dev/core/src/Materials/Textures/Loaders/ktxTextureLoader.ts
@@ -6,7 +6,6 @@ import type { IInternalTextureLoader } from "./internalTextureLoader";
 import { Logger } from "../../../Misc/logger";
 import { Constants } from "../../../Engines/constants";
 
-import "../../../Engines/Extensions/engine.cubeTexture";
 import type { Engine } from "core/Engines/engine";
 
 function mapSRGBToLinear(format: number): Nullable<number> {
