+++ packages/dev/core/src/Materials/Textures/rawTexture3D.ts
@@ -1,7 +1,6 @@
 import type { Scene } from "../../scene";
 import { Texture } from "./texture";
 import { Constants } from "../../Engines/constants";
-import "../../Engines/Extensions/engine.rawTexture";
 import type { Nullable } from "../../types";
 /**
  * Class used to store 3D textures containing user data
